Get Karma to work with Phantom, Systemjs and ES2016
Phantom is not yet able to run ES2015 and we are
compiling the frontend (for dev) to ES2016,
this was causing Karma, which uses Phantom, to fail.
We now use Babel with Karma to compile ES2016
to ES5 but the Karma babel plugin causes problems
with the base path defined to "base" by Karma, so we
mapped "/base" to "/" using the Karma config
"proxies" and remapped the files to be loaded
on karma-test-shim.js.